+++ src/components/Cart/CartRecipe.tsx
@@ -8,7 +8,7 @@ import {
 import { CART_MESSAGE } from '@/constants/message';
 import { FREE_SHIPPING_CONDITION } from '@/constants/system';
 import Info from '@/assets/Info.svg?react';
-import { recipeState } from '@/store/selectors/recipeSelector';
+import { orderRecipeState } from '@/store/selectors/orderRecipeSelector';
 import { selectedCouponListState } from '@/store/atoms';
 import styled from '@emotion/styled';
 import { theme } from '@/style/theme.style';
@@ -20,7 +20,8 @@ interface Props {
 }
 
 const CartRecipe = ({ isCoupon = false }: Props) => {
-  const { orderPrice, shippingFee, totalPrice } = useRecoilValue(recipeState);
+  const { orderPrice, shippingFee, totalPrice } =
+    useRecoilValue(orderRecipeState);
 
   const selectedCoupon = useRecoilValue(selectedCouponListState);
