+++ src/store/selectors/orderRecipeSelector.ts
@@ -1,12 +1,11 @@
-import { OrderedItem, Recipe } from '@/types/recipe.type';
-
 import { FREE_SHIPPING_CONDITION } from '@/constants/system';
+import { Recipe } from '@/types/recipe.type';
 import { orderItemState } from './orderItemSelector';
 import { selector } from 'recoil';
 import { shippingFeeState } from './shippingFeeSelector';
 
-export const recipeState = selector<Recipe>({
-  key: 'recipeState',
+export const orderRecipeState = selector<Recipe>({
+  key: 'orderRecipeState',
   get: ({ get }) => {
     const orderedList = get(orderItemState);
     const shippingAreaFee = get(shippingFeeState);
@@ -26,22 +25,5 @@ export const recipeState = selector<Recipe>({
       shippingFee,
       totalPrice,
     };
-  },
-});
-
-export const orderedItemQuantityState = selector<OrderedItem>({
-  key: 'orderedItemQuantityState',
-  get: ({ get }) => {
-    const orderedList = get(orderItemState);
-
-    return orderedList.reduce(
-      (acc, cur) => {
-        acc.itemCount++;
-        acc.totalQuantity += cur.quantity;
-
-        return acc;
-      },
-      { itemCount: 0, totalQuantity: 0 }
-    );
   },
 });
